move redeem ui to collator staking module
