+++ novawallet/Modules/Staking/Parachain/ParaStkRedeem/ParaStkRedeemViewFactory.swift
@@ -3,7 +3,7 @@ import SoraFoundation
 import SoraKeystore
 
 struct ParaStkRedeemViewFactory {
-    static func createView(for state: ParachainStakingSharedStateProtocol) -> ParaStkRedeemViewProtocol? {
+    static func createView(for state: ParachainStakingSharedStateProtocol) -> CollatorStakingRedeemViewProtocol? {
         let chainAsset = state.stakingOption.chainAsset
 
         guard
@@ -44,7 +44,10 @@ struct ParaStkRedeemViewFactory {
             logger: Logger.shared
         )
 
-        let view = ParaStkRedeemViewController(presenter: presenter, localizationManager: localizationManager)
+        let view = CollatorStakingRedeemViewController(
+            presenter: presenter,
+            localizationManager: localizationManager
+        )
 
         presenter.view = view
         interactor.presenter = presenter
